удалил mock data, поправил импорты в main, изменил successAlert
