+++ js/main.js
@@ -1,7 +1,3 @@
-import './data.js';
-import './util.js';
-import './card-generate.js';
-import './form-toggle.js';
-import './form-validating.js';
+import './form-validate.js';
 import './map.js';
-import './api.js';
+import './server.js';
